add "group" concept to performance timeline
There is a need for a common concept of an "event group" in the
Performance Timeline, which would allow the UA to associate one or more
distinct events (of same or different types) into a logical group.

Some examples of where this is necessary:
- Nav Timing: nav request -> one or more redirects -> destination
- Res Timing: cors preflight -> request; multi-part/resumed fetch
- Frame Timing: main frame -> one or more composite events
- ... and so on.

In each of the above cases the UA can assign a common group to indicate
that the events share a common parent. In turn, the developer can check
the group ID of each event and fetch other associated events if
necessary - e.g. follow a redirect chain; fetch render+composite events,
and so on.

This commit adds the following concepts:
- adds "group" identifier to PerformanceEntry
- adds getEntriesByGroup() to Performance Timeline