Add entry list to PerformanceObserverEntryList
This PR fixes https://github.com/w3c/performance-timeline/issues/167.
It does so by adding an associated list to PerformanceObserverEntryList and setting it upon construction. This way, the algorithms used in the getter methods actually make sense, as they require a raw list.

This PR also fixes a couple of links that are currently broken.