Check shouldAdd when appending to buffer
This PR renames the determine eligibility algorithm to a name that more clearly indicates the check done internally (checks whether the buffer is full). It also adds the shouldAdd check to right before adding the entry to the performance entry buffer, as this is now possible after https://github.com/w3c/timing-entrytypes-registry/pull/15. This is being done to allow Event Timing to only buffer entries of duration at least 104.