Add support for accessing cross-frame entries.
This adds a Document-Policy configuration point for documents, named
"share-performance-timeline-with", which acceps a set of origins. If it
is non-empty, then ancestor frames whose origins are contained in that
set can access the timeline entries of the child frame, either through
a PerformanceObserver, or by passing a boolean flag into the various
getEntries methods.