Add navigationId to PerformanceEntry (#192)
Add id and navigationId fields to PerformanceEntry. These are integers, used to uniquely identify timeline entries,
so that they can be referred to in other entries. The initial use case for this is to track navigation-like events within
a page's lifetime. For this, the navigationId field is used, in each entry, to refer to the most recent navigation (or
back-forward-cache restoration, or soft navigation) which had occurred in the page.

---------

Co-authored-by: Marcos Cáceres <marcos@marcosc.com>